requirements.txt
uvicorn[standard]
python-dotenv
passlib
fastapi
starlette
pydantic
bcrypt
pyjwt
sqlalchemy
alembic
psycopg2-binary